requirements/batch_ingestion.txt
-r base.txt
boto3==1.37.18
botocore==1.37.18
delta-spark==2.0.0
pyspark==3.2.4
py4j==0.10.9.5
apache-airflow>=2.7.3
apache-airflow-providers-docker>=3.7.0 